refactor(gateway): split cache.rs into cache/{mod,exact} module
Behavior-preserving refactor ahead of the semantic cache tier. cache.rs
becomes cache/mod.rs (shared infra: connection handling, replay protection,
raw KV) and the exact-match tier (cache_key/get/set + its tests) moves to
cache/exact.rs as a sibling impl block. The upcoming semantic.rs + embedder.rs
land alongside.

All 16 existing cache tests pass unchanged (now under cache::exact::tests and
cache::tests); clippy clean.
